Item11808: revert perltidy on all lib/CPAN and all plugins which are CoordinateWithAuthor, as requested by Micha
git-svn-id: http://svn.foswiki.org/trunk/StringifierContrib@14764 0b4bb1d4-4e5a-0410-9cc4-b2b747904278

+++ lib/Foswiki/Contrib/Stringifier/Plugins/PPT.pm
@@ -27,20 +27,24 @@ if (__PACKAGE__->_programExists($ppthtml)){
 
 sub stringForFile {
     my ($self, $filename) = @_;
+    my $tmp_file = tmpnam();
+    
+    return '' if (-f $tmp_file);
     
     # First I convert PPT to HTML
     my $cmd = $ppthtml . ' %FILENAME|F%';
     my ($output, $exit) = Foswiki::Sandbox->sysCommand($cmd, FILENAME => $filename);
     
     return '' unless ($exit == 0);
-
+    $output = $self->toUtf8($output);
+    
     # put the html into a temporary file
-    my ($fh, $tmp_file) = tmpnam();
-    print $fh $output;
+    open(TMPFILE, ">$tmp_file");
+    print TMPFILE $output;
+    close(TMPFILE);
 
     # use the HTML stringifier to convert HTML to TXT
-    my $stringifier = Foswiki::Contrib::Stringifier::Plugins::HTML->new();
-    my $text = $stringifier->stringForFile($tmp_file);
+    my $text = Foswiki::Contrib::Stringifier->stringFor($tmp_file);
 
     unlink($tmp_file);
